Refactor app structure and dynamic link handling
Moved dynamic link logic to a new core/dynamic_link/dynamic_link.dart file and updated MaterialApp setup to lib/my_app.dart for better modularity. Simplified home page navigation logic, removed unused code, and improved error handling. Bumped version to 7.7.1+15.
