+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: msbridge
 description: "MS Bridge: A cross-platform app for seamless content reading with online/offline access, Firebase integration, and a user-friendly interface."
-version: 7.7.0+14
+version: 7.7.1+15
 
 environment:
   sdk: ">=3.4.3 <4.0.0"
